requirements.txt
requests
python-dotenv
tqdm
numpy
pandas
tenacity
pytest